+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 ### Added
 - Github workflow for changelog verification ([#102](https://github.com/opensearch-project/opensearch-ruby/pull/102))
+- `opensearch-aws-sigv4` gem ([#104](https://github.com/opensearch-project/opensearch-ruby/pull/104))
 
 ### Changed